+++ components/progress/progress.tsx
@@ -1,16 +1,12 @@
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import Icon from '../icon';
-import { Circle } from 'rc-progress';
 import classNames from 'classnames';
+import Icon from '../icon';
 import { ConfigConsumer, ConfigConsumerProps } from '../config-provider';
 import { tuple } from '../_util/type';
-
-const statusColorMap: Record<string, string> = {
-  normal: '#108ee9',
-  exception: '#ff5500',
-  success: '#87d068',
-};
+import Line from './Line';
+import Circle from './Circle';
+import { validProgress } from './utils';
 
 const ProgressTypes = tuple('line', 'circle', 'dashboard');
 export type ProgressType = (typeof ProgressTypes)[number];
@@ -36,15 +32,6 @@ export interface ProgressProps {
   gapPosition?: 'top' | 'bottom' | 'left' | 'right';
   size?: ProgressSize;
 }
-
-const validProgress = (progress: number | undefined) => {
-  if (!progress || progress < 0) {
-    return 0;
-  } else if (progress > 100) {
-    return 100;
-  }
-  return progress;
-};
 
 export default class Progress extends React.Component<ProgressProps, {}> {
   static defaultProps = {
@@ -72,84 +59,24 @@ export default class Progress extends React.Component<ProgressProps, {}> {
     default: PropTypes.oneOf(['default', 'small']),
   };
 
-  renderLine(prefixCls: string, progressInfo: React.ReactNode) {
-    const { percent, successPercent, strokeWidth, size, strokeColor, strokeLinecap } = this.props;
-    const percentStyle = {
-      width: `${validProgress(percent)}%`,
-      height: strokeWidth || (size === 'small' ? 6 : 8),
-      background: strokeColor,
-      borderRadius: strokeLinecap === 'square' ? 0 : '100px',
-    };
-    const successPercentStyle = {
-      width: `${validProgress(successPercent)}%`,
-      height: strokeWidth || (size === 'small' ? 6 : 8),
-      borderRadius: strokeLinecap === 'square' ? 0 : '100px',
-    };
-    const successSegment =
-      successPercent !== undefined ? (
-        <div className={`${prefixCls}-success-bg`} style={successPercentStyle} />
-      ) : null;
-    return (
-      <div>
-        <div className={`${prefixCls}-outer`}>
-          <div className={`${prefixCls}-inner`}>
-            <div className={`${prefixCls}-bg`} style={percentStyle} />
-            {successSegment}
-          </div>
-        </div>
-        {progressInfo}
-      </div>
-    );
-  }
+  renderProcessInfo(prefixCls: string, progressStatus: (typeof ProgressStatuses)[number]) {
+    const { showInfo, format, type, percent, successPercent } = this.props;
+    if (!showInfo) return null;
 
-  renderCircle(prefixCls: string, progressInfo: React.ReactNode, progressStatus: string) {
-    const {
-      percent,
-      successPercent,
-      width,
-      strokeWidth,
-      strokeColor,
-      trailColor,
-      strokeLinecap,
-      gapPosition,
-      gapDegree,
-      type,
-    } = this.props;
-    const circleSize = width || 120;
-    const circleStyle = {
-      width: circleSize,
-      height: circleSize,
-      fontSize: circleSize * 0.15 + 6,
-    };
-    const circleWidth = strokeWidth || 6;
-    const gapPos = gapPosition || (type === 'dashboard' && 'bottom') || 'top';
-    const gapDeg = gapDegree || (type === 'dashboard' && 75);
-
-    // Merge values
-    let percents: number | number[] = validProgress(percent);
-    let strokeColors: string | string[] = strokeColor || statusColorMap[progressStatus];
-
-    if (successPercent) {
-      const successPercentVal = validProgress(successPercent);
-      percents = [successPercentVal, percents - successPercentVal];
-      strokeColors = [statusColorMap.success, strokeColors];
+    let text;
+    const textFormatter = format || (percentNumber => `${percentNumber}%`);
+    const iconType = type === 'circle' || type === 'dashboard' ? '' : '-circle';
+    if (format || (progressStatus !== 'exception' && progressStatus !== 'success')) {
+      text = textFormatter(validProgress(percent), validProgress(successPercent));
+    } else if (progressStatus === 'exception') {
+      text = <Icon type={`close${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
+    } else if (progressStatus === 'success') {
+      text = <Icon type={`check${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
     }
-
     return (
-      <div className={`${prefixCls}-inner`} style={circleStyle}>
-        <Circle
-          percent={percents}
-          strokeWidth={circleWidth}
-          trailWidth={circleWidth}
-          strokeColor={strokeColors}
-          strokeLinecap={strokeLinecap}
-          trailColor={trailColor}
-          prefixCls={prefixCls}
-          gapDegree={gapDeg}
-          gapPosition={gapPos}
-        />
-        {progressInfo}
-      </div>
+      <span className={`${prefixCls}-text`} title={typeof text === 'string' ? text : undefined}>
+        {text}
+      </span>
     );
   }
 
@@ -180,31 +107,23 @@ export default class Progress extends React.Component<ProgressProps, {}> {
       !('status' in props)
         ? 'success'
         : status || 'normal';
-    let progressInfo;
     let progress;
-    const textFormatter = format || (percentNumber => `${percentNumber}%`);
 
-    if (showInfo) {
-      let text;
-      const iconType = type === 'circle' || type === 'dashboard' ? '' : '-circle';
-      if (format || (progressStatus !== 'exception' && progressStatus !== 'success')) {
-        text = textFormatter(validProgress(percent), validProgress(successPercent));
-      } else if (progressStatus === 'exception') {
-        text = <Icon type={`close${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
-      } else if (progressStatus === 'success') {
-        text = <Icon type={`check${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
-      }
-      progressInfo = (
-        <span className={`${prefixCls}-text`} title={typeof text === 'string' ? text : undefined}>
-          {text}
-        </span>
-      );
-    }
+    const progressInfo = this.renderProcessInfo(prefixCls, progressStatus);
 
+    // Render progress shape
     if (type === 'line') {
-      progress = this.renderLine(prefixCls, progressInfo);
+      progress = (
+        <Line {...this.props} prefixCls={prefixCls}>
+          {progressInfo}
+        </Line>
+      );
     } else if (type === 'circle' || type === 'dashboard') {
-      progress = this.renderCircle(prefixCls, progressInfo, progressStatus);
+      progress = (
+        <Circle {...this.props} prefixCls={prefixCls} progressStatus={progressStatus}>
+          {progressInfo}
+        </Circle>
+      );
     }
 
     const classString = classNames(
